+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # merakitools - CHANGELOG
 ### v0.1.7
  - [new] Update network traffic analysis settings
+ - [new] List and create network webhook servers
 
 ### v0.1.6
  - [new] Update network settings